\name{bar-package}
\alias{bar-package}
\alias{bar}
\docType{package}
\title{
\packageTitle{bar}
}
\description{
\packageDescription{bar}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bar}
\packageIndices{bar}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{bar}

Maintainer: \packageMaintainer{bar}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
